Add KeysGenerator helper class to generate Keys.java
Currently the Keys.java is generated with Moxie: [1] Ant library. As
preparation for Bazel driven build, extract the MxKey moxie task in
standalone Java program and extend the interface to generate the key
in zip output file. To verify that generated file is the same compare
it with the one generated by mx:key moxie task.

Test Plan:

  $ bazel build :keys

[1] https://github.com/gitblit/moxie